# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API CRUD - Produtos e Categorias",
    "description": "Documentação da API para gerenciar produtos e categorias."
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["https"],
  "paths": {
    "/login": {
      "post": {
        "summary": "Realiza login",
        "description": "Retorna um token JWT após login bem-sucedido.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {"type": "string"},
                "password": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "200": {"description": "Login bem-sucedido"},
          "401": {"description": "Credenciais inválidas"}
        }
      }
    },
    "/categorias": {
      "post": {
        "summary": "Cria uma nova categoria",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {"type": "string"},
                "descricao": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "Categoria criada com sucesso"},
          "400": {"description": "Erro de validação"}
        }
      },
      "get": {
        "summary": "Lista todas as categorias",
        "responses": {
          "200": {"description": "Lista de categorias"}
        }
      }
    },
    "/produtos": {
      "post": {
        "summary": "Cria um novo produto",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {"type": "string"},
                "preco": {"type": "number"},
                "categoria_id": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "Produto criado com sucesso"},
          "400": {"description": "Erro de validação"}
        }
      },
      "get": {
        "summary": "Lista todos os produtos",
        "responses": {
          "200": {"description": "Lista de produtos"}
        }
      }
    }
  }
}